@@ -7,11 +7,11 @@
     "authors": [
         {
             "name": "Victor Toulouse",
-            "email": "toulouse@insa-toulouse.fr"
+            "email": "toulousevictor@gmail.com"
         }
     ],
     "require": {
-        "php": ">=5.6",
+        "php": ">=7",
         "doctrine/orm": "^2.5",
         "symfony/framework-bundle": "^2.8|^3.2",
         "symfony/form": "^2.8|^3.2",
