@@ -4,19 +4,15 @@ namespace PUGX\AutocompleterBundle\Loader;
 
 use Symfony\Component\Form\ChoiceList\ArrayChoiceList;
 use Symfony\Component\Form\ChoiceList\ChoiceListInterface;
-use \Symfony\Component\Form\ChoiceList\Loader\ChoiceLoaderInterface;
+use Symfony\Component\Form\ChoiceList\Loader\ChoiceLoaderInterface;
 
 /**
- * Class ChoiceLoader
- *
- * @package PUGX\AutocompleterBundle\Loader
+ * Class ChoiceLoader.
  */
 class ChoiceLoader implements ChoiceLoaderInterface
 {
-
     /** @var ChoiceListInterface */
     private $choiceList;
-
 
     /**
      * Loads a list of choices.
@@ -38,7 +34,7 @@ class ChoiceLoader implements ChoiceLoaderInterface
         }
 
         // if no values preset yet return empty list
-        $this->choiceList = new ArrayChoiceList(array(), $value);
+        $this->choiceList = new ArrayChoiceList([], $value);
 
         return $this->choiceList;
     }
@@ -53,22 +49,21 @@ class ChoiceLoader implements ChoiceLoaderInterface
      * The callable receives the choice as first and the array key as the second
      * argument.
      *
-     * @param string[] $values An array of choice values. Non-existing
+     * @param string[]      $values An array of choice values. Non-existing
      *                              values in this array are ignored
-     * @param null|callable $value The callable generating the choice values
+     * @param null|callable $value  The callable generating the choice values
      *
      * @return array An array of choices
      */
     public function loadChoicesForValues(array $values, $value = null)
     {
         // is called on form submit after loadValuesForChoices of form create and loadChoiceList of form view create
-        $choices = array();
+        $choices = [];
         foreach ($values as $key => $val) {
             // we use a DataTransformer, thus only plain values arrive as choices which can be used directly as value
             if (is_callable($value)) {
-                $choices[$key] = (string)call_user_func($value, $val, $key);
-            }
-            else {
+                $choices[$key] = (string) call_user_func($value, $val, $key);
+            } else {
                 $choices[$key] = $val;
             }
         }
@@ -89,18 +84,19 @@ class ChoiceLoader implements ChoiceLoaderInterface
      * The callable receives the choice as first and the array key as the second
      * argument.
      * 2017-01-25 vtoulouse: NOT TRUE! actually there is no second argument!
+     *
      * @see ArrayChoiceList::flatten()
      *
-     * @param array $choices An array of choices. Non-existing choices in
+     * @param array         $choices An array of choices. Non-existing choices in
      *                               this array are ignored
-     * @param null|callable $value The callable generating the choice values
+     * @param null|callable $value   The callable generating the choice values
      *
      * @return string[] An array of choice values
      */
     public function loadValuesForChoices(array $choices, $value = null)
     {
         // is called on form creat with $choices containing the preset of the bound entity
-        $values = array();
+        $values = [];
 
         // FIXME I had to check the layout of the given $choices array,
         // FIXME as for some reason it can come in three forms:
@@ -119,23 +115,19 @@ class ChoiceLoader implements ChoiceLoaderInterface
          */
         if (isset($choices[0]) && is_array($choices[0])) {
             foreach ($choices as $choice) {
-
                 if (is_callable($value)) {
-                    $values[key($choice)] = (string)call_user_func($value, current($choice), key($choice));
-                }
-                else {
+                    $values[key($choice)] = (string) call_user_func($value, current($choice), key($choice));
+                } else {
                     $values[key($choice)] = current($choice);
                 }
             }
-        } else
-
-        foreach ($choices as $key => $choice) {
-
-            if (is_callable($value)) {
-                $values[$key] = (string)call_user_func($value, $choice, $key);
-            }
-            else {
-                $values[$key] = $choice;
+        } else {
+            foreach ($choices as $key => $choice) {
+                if (is_callable($value)) {
+                    $values[$key] = (string) call_user_func($value, $choice, $key);
+                } else {
+                    $values[$key] = $choice;
+                }
             }
         }
 
