@@ -5,23 +5,21 @@ namespace PUGX\AutocompleterBundle\Form\Transformer;
 /**
  * The class objects are transformable in a numeric ID and a string representation
  * by implementing methods getId() and __toString().
- * This makes objects manageable by the CollectionToChoicesTransformer
- *
- * @package PUGX\AutocompleterBundle\Form\Transformer
+ * This makes objects manageable by the CollectionToChoicesTransformer.
  */
 interface TransformableInterface
 {
     /**
      * Returns a unique numeric identifier for the object.
      *
-     * @return integer
+     * @return int
      */
-    public function getId();
+    public function getId(): int;
 
     /**
      * Returns a string representation of the object.
      *
      * @return string
      */
-    public function __toString();
+    public function __toString(): string;
 }
